# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Viral Waitlist & Referral Marketing Platform API
  description: |
    REST API for the Viral Waitlist & Referral Marketing Platform.

    ## Authentication

    All API requests require authentication via:
    - **API Key:** Header `X-API-Key: sk_live_xxxxx` or `Authorization: Bearer sk_live_xxxxx`
    - **JWT Token:** For dashboard users, `Authorization: Bearer <jwt>`

    API Key Format:
    - Test: `sk_test_` (test mode)
    - Live: `sk_live_` (production)

    ## Rate Limiting

    Rate limits by tier (per minute):
    - **Free:** 60 requests
    - **Starter:** 60 requests
    - **Pro:** 180 requests
    - **Enterprise:** 600+ requests (custom)

    Response headers include:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the rate limit resets
    - `X-Request-ID`: Unique request identifier for debugging

    ## Error Handling

    All error responses follow a consistent format with both a human-readable message and a machine-readable error code:

    ```json
    {
      "error": "User-friendly error message",
      "code": "ERROR_CODE"
    }
    ```

    Error codes use `UPPER_SNAKE_CASE` format for easy programmatic handling.

    ### Common Error Codes

    - `INVALID_INPUT` - Validation error or malformed request
    - `UNAUTHORIZED` - Missing or invalid authentication
    - `FORBIDDEN` - Insufficient permissions
...

...

...
servers:
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /api/v1/campaigns:
    post:
      tags:
        - Campaigns
      summary: Create a new campaign
      description: Creates a new waitlist or referral campaign
      operationId: createCampaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
      responses:
        '201':
          description: Campaign created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - Campaigns
      summary: List all campaigns
      description: Retrieve a list of all campaigns for the authenticated account
      operationId: listCampaigns
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          description: Filter by campaign status
          schema:
            type: string
            enum: [draft, active, paused, completed]
        - name: type
          in: query
          description: Filter by campaign type
          schema:
            type: string
            enum: [waitlist, referral, contest]
      responses:
        '200':
          description: List of campaigns
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListCampaignsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/campaigns/{campaign_id}:
    parameters:
      - $ref: '#/components/parameters/CampaignIdParam'

    get:
      tags:
        - Campaigns
      summary: Get campaign by ID
      description: Retrieve detailed information about a specific campaign
      operationId: getCampaign
      responses:
        '200':
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags:
        - Campaigns
      summary: Update campaign
      description: Update an existing campaign's configuration
      operationId: updateCampaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCampaignRequest'
      responses:
        '200':
          description: Campaign updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Campaigns
      summary: Delete campaign
      description: Soft delete a campaign (can be recovered)
      operationId: deleteCampaign
      responses:
        '204':
          description: Campaign deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/campaigns/{campaign_id}/status:
    patch:
      tags:
        - Campaigns
      summary: Update campaign status
      description: Update the status of a campaign (draft, active, paused, completed)
      operationId: updateCampaignStatus
      parameters:
        - $ref: '#/components/parameters/CampaignIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [draft, active, paused, completed]
                  description: New status for the campaign
      responses:
        '200':
          description: Campaign status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

  # ==================== WAITLIST USERS ====================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key authentication (sk_live_* or sk_test_*)

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication for dashboard users
  parameters:
    CampaignIdParam:
      name: campaign_id
      in: path
      required: true
      description: Campaign unique identifier
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validation:
              value:
                error: "Invalid request format. Please check your JSON syntax."
                code: "INVALID_INPUT"
            missing_field:
              value:
                error: "Required field is missing"
                code: "INVALID_INPUT"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authorization token is missing or invalid"
            code: "UNAUTHORIZED"
...
    NotFound:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            campaign:
              value:
                error: "Campaign not found"
                code: "CAMPAIGN_NOT_FOUND"
            user:
              value:
                error: "User not found"
                code: "USER_NOT_FOUND"
    Conflict:
      description: Conflict - Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            email_exists:
              value:
                error: "Email already exists"
                code: "EMAIL_EXISTS"
            user_exists:
              value:
                error: "User already exists in this campaign"
                code: "USER_EXISTS"
    TooManyRequests:
      description: Too many requests - Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Maximum requests per window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Unix timestamp when rate limit resets
          schema:
            type: integer
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Rate limit exceeded. Please try again later."
            code: "RATE_LIMIT_EXCEEDED"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "An internal server error occurred. Please try again later."
            code: "INTERNAL_ERROR"
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: User-friendly error message
        code:
          type: string
          description: Machine-readable error code in UPPER_SNAKE_CASE format
          example: "INVALID_INPUT"
      example:
        error: "Invalid request format"
        code: "INVALID_INPUT"

    # ==================== CORE MODELS ====================
    Campaign:
      type: object
      required:
        - id
        - account_id
        - name
        - slug
        - status
        - type
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 255
        slug:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        status:
          type: string
          enum: [draft, active, paused, completed]
        type:
          type: string
          enum: [waitlist, referral, contest]
        launch_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        form_config:
          $ref: '#/components/schemas/FormConfig'
        referral_config:
          $ref: '#/components/schemas/ReferralConfig'
        email_config:
          $ref: '#/components/schemas/EmailConfig'
        branding_config:
          $ref: '#/components/schemas/BrandingConfig'
        privacy_policy_url:
          type: string
          format: uri
        terms_url:
          type: string
          format: uri
        max_signups:
          type: integer
        total_signups:
          type: integer
        total_verified:
          type: integer
        total_referrals:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    FormConfig:
      type: object
      properties:
        fields:
          type: array
          items:
            $ref: '#/components/schemas/FormField'
        captcha_enabled:
          type: boolean
          default: false
        double_opt_in:
          type: boolean
          default: true
        custom_css:
          type: string
    FormField:
      type: object
      required:
        - name
        - type
        - label
      properties:
        name:
          type: string
        type:
          type: string
          enum: [email, text, select, checkbox, textarea, number]
        label:
          type: string
        placeholder:
          type: string
        required:
          type: boolean
          default: false
        options:
          type: array
          items:
            type: string
          description: Options for select fields
        validation:
          type: string
          description: Regex pattern for validation
    ReferralConfig:
      type: object
      properties:
        enabled:
          type: boolean
          default: true
        points_per_referral:
          type: integer
          default: 1
        verified_only:
          type: boolean
          default: true
        sharing_channels:
          type: array
          items:
            type: string
            enum: [email, twitter, facebook, linkedin, whatsapp]
        custom_share_messages:
          type: object
          additionalProperties:
            type: string
    EmailConfig:
      type: object
      properties:
        from_name:
          type: string
        from_email:
          type: string
          format: email
        reply_to:
          type: string
          format: email
        verification_required:
          type: boolean
          default: true
    BrandingConfig:
      type: object
      properties:
        logo_url:
          type: string
          format: uri
        primary_color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          default: '#2563EB'
        font_family:
          type: string
          default: 'Inter'
        custom_domain:
          type: string
          format: hostname
...
    CreateCampaignRequest:
      type: object
      required:
        - name
        - slug
        - type
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        slug:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '^[a-z0-9-]+$'
        description:
          type: string
        type:
          type: string
          enum: [waitlist, referral, contest]
        form_config:
          $ref: '#/components/schemas/FormConfig'
        referral_config:
          $ref: '#/components/schemas/ReferralConfig'
        email_config:
          $ref: '#/components/schemas/EmailConfig'
        branding_config:
          $ref: '#/components/schemas/BrandingConfig'
    UpdateCampaignRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        launch_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        form_config:
          $ref: '#/components/schemas/FormConfig'
        referral_config:
          $ref: '#/components/schemas/ReferralConfig'
        email_config:
          $ref: '#/components/schemas/EmailConfig'
        branding_config:
          $ref: '#/components/schemas/BrandingConfig'
        max_signups:
          type: integer
...
    ListCampaignsResponse:
      type: object
      properties:
        campaigns:
          type: array
          description: Array of campaigns. Returns empty array [] if no campaigns found, never null.
          items:
            $ref: '#/components/schemas/Campaign'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        next_cursor:
          type: string
          description: Cursor for next page
        has_more:
          type: boolean
        total_count:
          type: integer